remove unused imports and rename package
