+++ src/main/java/com/saayam/volunteer/VolunteerApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class VolunteerApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(VolunteerApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(VolunteerApplication.class, args);
+    }
 
 }
