+++ src/main/java/com/saayam/volunteer/dto/CreateUserResponse.java
@@ -1,4 +1,4 @@
-package com.saayam.volunteer.domain;
+package com.saayam.volunteer.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
