+++ src/main/java/com/saayam/volunteer/handler/CreateUserHandler.java
@@ -9,8 +9,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.saayam.volunteer.VolunteerApplication;
-import com.saayam.volunteer.domain.CreateUserRequest;
-import com.saayam.volunteer.domain.CreateUserResponse;
+import com.saayam.volunteer.dto.CreateUserRequest;
+import com.saayam.volunteer.dto.CreateUserResponse;
 import com.saayam.volunteer.service.UserService;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
