+++ src/main/java/com/saayam/volunteer/handler/UpdateUserProfileHandler.java
@@ -8,8 +8,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.saayam.volunteer.VolunteerApplication;
-import com.saayam.volunteer.domain.UpdateUserProfileRequest;
-import com.saayam.volunteer.domain.UserProfileResponse;
+import com.saayam.volunteer.dto.UpdateUserProfileRequest;
+import com.saayam.volunteer.dto.UserProfileResponse;
 import com.saayam.volunteer.service.UserService;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
