+++ src/main/java/com/saayam/volunteer/model/State.java
@@ -17,7 +17,6 @@ import lombok.NoArgsConstructor;
 
 import java.time.ZonedDateTime;
 import java.util.Objects;
-import java.util.UUID;
 
 @Entity
 @Data
