+++ src/main/java/com/saayam/volunteer/model/UserCategory.java
@@ -12,7 +12,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.ZonedDateTime;
-import java.util.UUID;
 
 @Entity
 @Data
