+++ src/main/java/com/saayam/volunteer/model/UserStatus.java
@@ -12,8 +12,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.ZonedDateTime;
-import java.util.Objects;
-import java.util.UUID;
 
 @Entity
 @Data
