+++ src/main/java/com/saayam/volunteer/repository/CountryRepository.java
@@ -4,8 +4,6 @@ import com.saayam.volunteer.model.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
 public interface CountryRepository extends JpaRepository<Country, Integer> {
 }
