+++ src/main/java/com/saayam/volunteer/repository/UserStatusRepository.java
@@ -4,8 +4,6 @@ import com.saayam.volunteer.model.UserStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
 public interface UserStatusRepository extends JpaRepository<UserStatus, Integer> {
 }
